merge

+++ ruhuna-schedule-ease/app/Http/Controllers/AnnouncementController.php
@@ -8,21 +8,24 @@ use Inertia\Inertia;
 
 class AnnouncementController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        $this->authorize('read_announcement', $request->user());
         $announcements = Announcement::all();
         return Inertia::render('Announcements/Index', [
             'announcements' => $announcements
         ]);
     }
 
-    public function create()
+    public function create(Request $request)
     {
+        $this->authorize('create_announcement', $request->user());
         return Inertia::render('Announcements/Create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create_announcement', $request->user());
         $request->validate([
             'title' => 'required|string|max:255',
             'body' => 'required|string',
@@ -32,8 +35,9 @@ class AnnouncementController extends Controller
         return redirect()->route('announcements.index');
     }
 
-    public function edit(Announcement $announcement)
+    public function edit(Announcement $announcement,Request $request)
     {
+        $this->authorize('update_announcement', $request->user());
         return Inertia::render('Announcements/Edit', [
             'announcement' => $announcement
         ]);
@@ -41,6 +45,7 @@ class AnnouncementController extends Controller
 
     public function update(Request $request, Announcement $announcement)
     {
+        $this->authorize('update_announcement', $request->user());
         $request->validate([
             'title' => 'required|string|max:255',
             'body' => 'required|string',
@@ -50,8 +55,9 @@ class AnnouncementController extends Controller
         return redirect()->route('announcements.index');
     }
 
-    public function destroy(Announcement $announcement)
+    public function destroy(Announcement $announcement,Request $request)
     {
+        $this->authorize('delete_announcement', $request->user());
         $announcement->delete();
         return redirect()->route('announcements.index');
     }
